refactor: share benchmark isolation harness
