+++ tests/test_mbpp.py
@@ -179,75 +179,6 @@ def test_incomplete_solution_gets_dense_partial_score(tmp_path, trusted_grader):
     assert "2/3" in result.detail
 
 
-def test_agent_cannot_replace_the_held_out_grader(tmp_path, trusted_grader):
-    task, ws, _ = _seed_task(tmp_path)
-    (ws / "_grade.py").write_text(
-        'print("{\\"total\\": 3, \\"passed\\": 3}")\n', encoding="utf-8"
-    )
-    result = task.grade(ws, sandbox=trusted_grader)
-    assert result.score == 0.0 and not result.passed
-    assert not (ws / "_grade.py").exists()
-
-
-def test_candidate_cannot_disable_assertions_by_patching_exec(tmp_path, trusted_grader):
-    task, ws, _ = _seed_task(tmp_path)
-    (ws / "solution.py").write_text(
-        "import builtins\nbuiltins.exec = lambda *args, **kwargs: None\n", encoding="utf-8"
-    )
-    result = task.grade(ws, sandbox=trusted_grader)
-    assert result.score == 0.0 and not result.passed
-
-
-def test_candidate_cannot_reach_trusted_control_through_frames(tmp_path, trusted_grader):
-    task, ws, _ = _seed_task(tmp_path)
-    (ws / "solution.py").write_text(
-        "import sys\n"
-        "caller = sys._getframe(1).f_globals\n"
-        "caller['trusted_exec'] = lambda *args, **kwargs: None\n",
-        encoding="utf-8",
-    )
-    result = task.grade(ws, sandbox=trusted_grader)
-    assert result.score == 0.0 and not result.passed
-
-
-def test_task_local_module_cannot_shadow_trusted_grader_imports(tmp_path, trusted_grader):
-    task, ws, _ = _seed_task(tmp_path)
-    (ws / "base64.py").write_text(
-        "import os\nimport sys\n"
-        "frame = sys._getframe()\n"
-        "while frame is not None:\n"
-        "    scope = frame.f_globals\n"
-        "    if 'REPORT_PREFIX' in scope and 'TESTS' in scope:\n"
-        "        total = len(scope['TESTS'])\n"
-        "        sys.stdout.write(scope['REPORT_PREFIX'] + f'{total}/{total}\\n')\n"
-        "        sys.stdout.flush()\n"
-        "        os._exit(0)\n"
-        "    frame = frame.f_back\n"
-        "raise ImportError('trusted parent not found')\n",
-        encoding="utf-8",
-    )
-    result = task.grade(ws, sandbox=trusted_grader)
-    assert result.score == 0.0 and not result.passed
-
-
-def test_candidate_cannot_forge_a_late_json_report(tmp_path, trusted_grader):
-    task, ws, _ = _seed_task(tmp_path)
-    (ws / "solution.py").write_text(
-        "import atexit\n"
-        "atexit.register(lambda: print('{\"total\": 3, \"passed\": 3}'))\n",
-        encoding="utf-8",
-    )
-    result = task.grade(ws, sandbox=trusted_grader)
-    assert result.score == 0.0 and not result.passed
-
-
-def test_candidate_early_exit_cannot_pass(tmp_path, trusted_grader):
-    task, ws, _ = _seed_task(tmp_path)
-    (ws / "solution.py").write_text("import os\nos._exit(0)\n", encoding="utf-8")
-    result = task.grade(ws, sandbox=trusted_grader)
-    assert result.score == 0.0 and not result.passed
-
-
 def test_unavailable_sandbox_returns_legible_zero(tmp_path):
     class UnavailableSandbox:
         def run(self, *args, **kwargs):
@@ -257,32 +188,6 @@ def test_unavailable_sandbox_returns_legible_zero(tmp_path):
     result = task.grade(ws, sandbox=UnavailableSandbox())
     assert result.score == 0.0 and not result.passed
     assert "secure grader unavailable" in result.detail
-
-
-def test_malformed_grader_report_returns_zero(tmp_path):
-    import subprocess
-
-    class MalformedSandbox:
-        def run(self, *args, **kwargs):
-            return subprocess.CompletedProcess(["python", "_grade.py"], 0, "noise\n", "")
-
-    task, ws, _ = _seed_task(tmp_path)
-    result = task.grade(ws, sandbox=MalformedSandbox())
-    assert result.score == 0.0 and not result.passed
-    assert "no report" in result.detail
-
-
-def test_none_grader_streams_return_zero_instead_of_raising(tmp_path):
-    import subprocess
-
-    class NoneStreamsSandbox:
-        def run(self, *args, **kwargs):
-            return subprocess.CompletedProcess(["python", "_grade.py"], 0, None, None)
-
-    task, ws, _ = _seed_task(tmp_path)
-    result = task.grade(ws, sandbox=NoneStreamsSandbox())
-    assert result.score == 0.0 and not result.passed
-    assert "no report" in result.detail
 
 
 def test_cli_resolves_mbpp_task(tmp_path):
